# Conflicts:
#	features/top_ads/top_ads_dashboard/src/main/java/com/tokopedia/topads/dashboard/view/presenter/TopAdsDashboardPresenter.kt

+++ features/top_ads/top_ads_headline/src/main/java/com/tokopedia/top_ads_headline/data/TopAdsManageHeadlineInput.kt
@@ -71,7 +71,7 @@ data class TopAdsManageHeadlineInput(
                         @SerializedName("id")
                         var id: String = "0",
                         @SerializedName("priceBid")
-                        var priceBid: Int = 0,
+                        var priceBid: Long = 0,
                         @SerializedName("status")
                         var status: String = "",
                         @SerializedName("tag")
